Starting masters degree

+++ addLight.m
@@ -1,6 +1,6 @@
 function [result] = addLight(img)
     min_val = min(min(img));
-    result = double(img) - min_val;
+    result = img - min_val;
     delta = max(max(result));
     result = uint8(result*255/delta);
 end
